archive/rpgCore/archive/legacy_refactor_2026/requirements_solid.txt
# SOLID Architecture Asset Ingestor Requirements
# Production-ready dependencies with version pinning

# Core dependencies
pydantic==2.6.0
loguru==0.7.2
rich==13.7.0

# Image processing
Pillow==10.2.0
numpy==1.26.4

# Testing and development
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
black==24.1.1
isort==5.13.2
mypy==1.8.0

# Type checking and validation
types-Pillow==10.2.0.20240106

# Optional performance monitoring (uncomment if needed)
# psutil==5.9.8
# memory-profiler==0.61.0
